fix(*): Fixed bug cannot unmarshal number
Bugs fixed
json: cannot unmarshal number -1703168784 into Go struct field Tx.ver of type uint64
json: cannot unmarshal number -1 into Go struct field Block.fee of type uint64
json: cannot unmarshal number 1.51806521E12 into Go struct field Stats.timestamp of type uint64

+++ chart.go
@@ -44,7 +44,7 @@ type Stats struct {
 	TotalBtcSent                  int64   `json:"total_btc_sent"`
 	TradeVolumeBtc                float64 `json:"trade_volume_btc"`
 	TradeVolumeUsd                float64 `json:"trade_volume_usd"`
-	Timestamp                     uint64  `json:"timestamp"`
+	Timestamp                     float64 `json:"timestamp"`
 }
 
 // GetStats is a special mechanism to obtain data behind Blockchain's stats
